localization: Change chat_id type to long (int64)

+++ sophie/components/localization/db/locale.py
@@ -28,7 +28,7 @@ col_validation = {
             "required": ["chat_id", "locale_code"],
             "properties": {
                 "chat_id": {
-                    "bsonType": "int"
+                    "bsonType": "long"
                 },
                 "locale_code": {
                     "bsonType": "string",
